requirements.txt
# Core ML frameworks
torch>=2.0.0
transformers>=4.36.0

# PEFT and efficient fine-tuning
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0

# Hugging Face Hub
huggingface-hub>=0.19.0

# Data processing
datasets>=2.14.0
pandas>=1.3.0
numpy>=1.21.0

# Progress and utilities
tqdm>=4.62.0
pyyaml>=5.4.0

# Tokenizers
sentencepiece>=0.1.99
protobuf>=3.20.0

# Logging and monitoring
tensorboard>=2.14.0

# Evaluation and analysis (optional)
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0

# LangChain for tool calling
langchain>=0.1.0
langchain-community>=0.0.20
pydantic>=2.0.0


